feat: extract sex generator from name.gender to name.sex
