--- src/locales/en/name/binary_gender.ts
@@ -1,1 +1,0 @@
-export default ['Female', 'Male'];
