+++ src/locales/en/name/sex.ts
@@ -1,0 +1,1 @@
+export default ['female', 'male', 'non-binary'];
