+++ src/locales/fr_CH/name/index.ts
@@ -3,23 +3,23 @@
  * Run 'pnpm run generate:locales' to update.
  */
 import type { NameDefinitions } from '../../..';
-import binary_gender from './binary_gender';
 import female_first_name from './female_first_name';
 import first_name from './first_name';
 import last_name from './last_name';
 import male_first_name from './male_first_name';
 import name_ from './name';
 import prefix from './prefix';
+import sex from './sex';
 import title from './title';
 
 const name: NameDefinitions = {
-  binary_gender,
   female_first_name,
   first_name,
   last_name,
   male_first_name,
   name: name_,
   prefix,
+  sex,
   title,
 };
