+++ src/locales/ur/name/sex.ts
